requirements.txt
uvicorn
fastapi
sqlalchemy
psycopg2
python-multipart
pyjwt[crypto]
pytest
pytest-dotenv
httpx